ocr-server/requirements.txt
flask
pillow
opencv-python-headless
numpy
easyocr
transformers
torch
